pyproject.toml
[project]
name = "ai-doctor"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "litellm>=1.83.14",
    "lxml>=6.1.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "simple-icd-10-cm>=1.5.0",
    "sse-starlette>=3.4.1",
    "uvicorn[standard]>=0.46.0",
]
